LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-anchor-idl-property.tentative-expected.txt

This is an anchor button  This button invokes the popover but isn't an anchor
button Anchored div

FAIL popover anchorElement IDL property returns the anchor element assert_equals: expected (object) Element node <button id="b1">This is an anchor button</button> but got (undefined) undefined
FAIL popover anchorElement is settable assert_equals: expected (object) Element node <button id="b1">This is an anchor button</button> but got (undefined) undefined
FAIL anchorElement affects anchor positioning assert_equals: expected (object) null but got (undefined) undefined
